change: change in button layout and name
