+++ src/entrypoint/frontend/lib/app/login/ui/widgets/custom_button.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+import '../../../../core/theme/mh_colors.dart';
+
+class CustomButton extends StatelessWidget {
+  const CustomButton({
+    super.key,
+    required this.text,
+    required this.onTap,
+  });
+
+  final String text;
+  final Function() onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: onTap,
+      borderRadius: BorderRadius.circular(24),
+      child: Ink(
+        height: 45,
+        width: double.infinity,
+        decoration: BoxDecoration(
+          color: MHColors.blue,
+          borderRadius: BorderRadius.circular(24),
+        ),
+        child: Center(
+          child: Text(
+            text,
+            style: const TextStyle(
+              color: MHColors.white,
+              fontSize: 16,
+              fontWeight: FontWeight.w600,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
