--- src/entrypoint/frontend/lib/app/login/ui/widgets/custom_elevated_button.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../../../core/theme/mh_colors.dart';
-
-class CustomElevatedButton extends StatelessWidget {
-  const CustomElevatedButton({
-    super.key,
-    required this.text,
-    required this.onPressed,
-  });
-
-  final String text;
-  final Function() onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: onPressed,
-      style: ButtonStyle(
-        backgroundColor: WidgetStateProperty.all(
-          MHColors.purple,
-        ),
-        shape: WidgetStateProperty.all(
-          RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20),
-          ),
-        ),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 6, vertical: 6),
-        child: Text(
-          text,
-          style: const TextStyle(
-            color: MHColors.white,
-          ),
-        ),
-      ),
-    );
-  }
-}
